New data quality control tool. Talked about it a number of times. DQ and engineering are in real need of it for investigating some data quality issue with the data pool at hand. Borrowed code from all over the place. Nice exercise for testing the robustness of the existing. Some code could benefit from more sophisticated error handling.
In a nutshell, 'data_qc' compares the actual file sizes of CEOS data
sets with estimates derived from the data. 
For the image data file estimate we read the image file descriptor
record.
For the leader file we read all available CEOS records and add up the
record lengths.
The ALOS trailer file contains low resolution image data. We don't
want to extract it but need to estimate its size. Hence, the extension
of reading the trailer file descriptor record.
For the volume file we use the same approach as the leader file.
The ALOS workreport is a special case. Only checking for existence and
for zero file size.

This definitely works for ALOS Palsar level zero data, the type of
data set that needs to immediately investigated. However, it should
work for the majority of the other CEOS flavors. Haven't tested that yet.