Fix fileSize() to work even when stat() fails, by falling back to using a bunch of fread()s.
But I don't think we need this, actually.  The old fileSize() code
would check stat() on all files in the directory, instead of just the
one we are interested in -- and, in particular, trying to check "."
and ".." would sometimes fail.  Fixing that got rid of the error, but
I have left the fallback code in there anyway.