Changes as requested by DQ.
The verbose mode is now default.

Added a check for image file size versus file size according to JAXA
spec. Anything outside the spec sets the status to FAILED. This can be
overwritten with the new -ignore_spec option.

Added an essential flag to only check image and leader file. That does
not complain about a missing workreport file or such thing but checks
the integrity of the files that we care about for ingest.