Added ability to compare images containing complex number values.
Found bug in get_maxval() ...a key element for
calculating appropriate PSNR limits.  I'm surprised
it has worked all this time, but now ...it means
that I'll need to keep an eye on autoregress results
to make sure that the PSNR calculation is still 
sensitive enough to catch differences.  I'll be
doing some more T&D in the mean time...