Revamped the AirSAR ingest after talking to Bruce Chapman. Apparently the most trusted metadata (peg point and offsets) are in the DEM headers. This is for interferometric data only. Haven't touched polarimetric yet.
Now we are looking for the DEM header of the C-band data first. If
that is not available, we look for the L-band DEM. If that is not
available we error out.

Found an error in the metadata parser that set the lat/lon of the DEM
header peg point to zero. Seems to populate fine now.

This reshuffling resulted now in bootstrapping the amplitude and
coherence image to the same geometry as the DEM. Commented the
fudge_airsar_params function out for the moment.

Ready for Rick's testing once the MapReady bug 1095 is fixed.