Error out when files are not successfully added.
We used to have an error like this, but it got annoying when users
would select a bunch of D&L files, and the L files would fail to add.
Now we can add it back, though, because we have file filters (so users
can get rid of the L files), and we also handle metadata files
specially so they are silently ignored as we want, separately.

Fixes mantis 1110