Some rudimentary TerraSAR-X support.
The import knows something about the TERRASAR format. It expects to
get the XML metadata file (that has information about the location of
the actual data in it - hey, no messing with the directory,
understand?) as the input file.

I had only complex COSAR data (SLC), so that is what is currently
implemented. It errors out for everything else. We generate a
two-banded image (ASF internal style) and some very basic
metadata. Has enough information in there for asf_view, resample and
asf_export to work.

Next step is going to be the libxml2 library. Needs some deep look
into the spec to map the XML metadata to our metadata structure.