A problem that the planner has is the use of a reference time/orbit number to calculate the current orbit number.  This reference gets less and less accurate as it gets older, leading to incorrect path numbers when planning near the equator (and, eventually, not so near the equator).
So, I've put in some code to recalibrate the reference time using
the latest TLE info.  Whenver the TLE is updated, the reference
should be recalibrated.  Right now, we don't support automatic
updating of the TLE, so the calibration stuff that I've just added
here is a command-line option... but really, we should add a button
that updates the TLE, and automatically recalibrates.

For now, this will eliminate the planning errors near the equator
but in a couple months we'll be off again.