Checking in the modifications to terrain correction to support using the original ground range DEM during radiometric correction.  Had to get reskew_dem to dump this out, to make sure it is aligned the same as the slant range one, and then pass this along to deskew_dem.
We could use the new gr DEM in deskew_dem for geometric correction
as well, however this looks somewhat worse to me, so I left it so
that the back-converted-from-slant-range DEM is used for geometric
correction as before, and the new DEM is only used for the
radiometric calculations, but this is something we should probably
look into further.

This should eliminate the odd terracing that we've been seeing in
radiometrically corrected data.