Changes required to ensure that all the CSTARS data ingest.
Tested it with the all the data sets in testsuite 3.

'cstars_swb' is a bad data set since it does not have a dataset summary
record. Probably a trailer file missing. Need to see whether I have
the original in some backup.

Next step is looking for geolocation issues and fix those.