Change extract_{double,int,string}_options so they do not change the value of the variable that will contain the option's value, if the option is not present.
This allows code like this to work:

  double offset=0;
  extract_double_options(&argc, &argv, &offset, "-offset", "-o", NULL);
  asfPrintStatus("Using offset: %f\n", offset);

Where the variable is first set to a default value, and if the user
does not use the "offset" option, the default value will still be
there (instead of being set to NAN, as before).

To detect if the command-line option was given or not, the return
value of the extract function should be used.