A couple of fixes.
Mantis 1292, 1299: We recently changed how the ALOS geocoding is done,
instead of using the reverse transform we use the forward transform
iteratively in the implementation of meta_get_lineSamp(), like we do
for ERS/RSAT.  This did improve geolocation for Prism/Avnir, but has
actually made things much worse for Palsar.  So, switched back to the
old implementation (which is faster, anyway) for Palsar only.

Mantis 1293: Had to tighten up the tolerance on the iterative version
of meta_get_lineSamp() in order to pass the symmetry test for this
data.  However, this makes geocoding quite a bit slower -- so, we now
use the normal tolerane first, and do the symmetry test, and if it
fails we tighten it, and retry the symmetry test.  If we fail to
pass after a few tightenings we give up as before (though the user
may still override with -force), but if we do pass we proceed to
geocode using the found tolerance in meta_get_lineSamp().