Implemented the basic support for the use of map projected coordinates in shapefiles.
Needed to shift a few things around to allow the parsing of projection
information into a function that writes an ESRI projection file.

Currently, I have implemented a method for leader/metadata files that
convert the projection information from a metadata structure. The hook
for reading the information from a projection file is there. Haven't
implemented the projection file part yet that can be used for other
conversions such csv2shape and the like.

This whole thing needs some more testing to iron out all the issues.