Added four more image_data_types to the mix: POLARIMETRIC_SEGMENTATION POLARIMETRIC_DECOMPOSITION POLARIMETRIC_PARAMETERS POLARIMETRIC_MATRIX
The polarimetric segmentation (just staying in PolSARpro lingo here)
type covers all the classification scheme out there. This is for
polarimetric results that are single band and need to have a LUT
applied.

The polarimetric decomposition covers all the multiband (anything from
two to four bands actually) polarimetric outputs from the Pauli,
Freeman-Durden and alike processing schemes. They need to be assigned
color bands. For the next point delivery we will probably need to
treat them as separate files. The way to combine them is using the
command line tool 'adjust_bands'.

The polarimetric parameters type consists of 'entropy', 'anisotropy' and
'alpha' as well as eigenvalues/eigenvectors.

The polarimetrc matrix types is meant for matrix elements such as T11
etc. (from T3).

Not sure whether we need to the distinction between the last two
groups but I did not want to touch the code again, in case we need it.