Added the equi-rectangular map projection to asf_geocode. Will only use it for generating overlay images, so I did not add anything to the description of the command line.
I don't think we want to add this projection to our regular mix in
MapReady. Needed to force the geocoding through in all test
cases. Makes kind of sense considering that the origin of the
projection is where the Greenwich meridian meets the equator. My test
cases were Cook Inlet, Argentina and Hawaii (kind of far away from that).