Added transparency to PNG imagery.
First defined a format 'PNG_ALPHA' that sets a black background to
transparent. Works fine, even of AVNIR color imagery.

However, trying that for the Google Earth overlay, I had to realize
that GE does not support this type of transparency. Regardless whether
the image is greyscale or not, it wants a full blown alpha channel in
a 32-bit RGBA image. Talk about waste of space!

Did not want to mess around with the code to much, so I defined
another format: PNG_GE.

asf_mapready and asf_export can both handle these new formats.