Sorted out the rest of the PolSARPro handling. No the ENVI headers look identical to the ones that GDAL produces. Furthermore, the projection block in the metadata looks the same when ingesting the ENVI headers into ASF internal format again.
Still need to verify that the GeoTIFF files are working out the same
way, regardless of the processing flow. The regular processing flow
to GeoTIFF within MapReady needs to be the same as processing to
geocoded PolSARPro format files (to be worked on within PolSARPro and
reingested into MapReady to be exported into GeoTIFF).