Finally got the KML overlays properly hooked up with the GUI.
Needed some more work with the status updates. Had to introduce two
new functions 'status_on' and 'status_off' that set a flag used by
'update_status' to verify whether the status line in the GUI needs
updating. This was required as I am calling asf_mapready recursively.

Cleaned up the "Successful completion" statements in the related tools
as well to sort of the remaining confusion.