Preliminary check in for the generic metadata ingest tool, called asf_fgdc_meta.
Needed to shift around a few things in asf_meta_raw, so I got enough
metadata without actually ingesting the data.

Put functionality in place to write the metadata into an FGDC
compliant XML format. It actually validates with the USGS metadata
validation tool. There are few parameters left to parse from the
configuration file (abstract, purpose, keywords and lineage) but I
have all the metadata that I need to ingest into the database.

There are quite a number of map projection issues that are not passed
along correctly that I would like to fix. These are technically
MapReady bugs. I guess these are on as time permits basis.