This is a preliminary check in.
Implemented the tool for ingesting FGDC compliant metadata and KML
files that contain information about image boundary and image center.

The XML parsing into a data structure is done. Now I need to feed that
structure into the database. That will need some Oracle precompiler
stuff that only exists on a machine with a database on.