Fixed a small bug in the metadata parser.
Adapted the export to handle an InSAR stack. Currently, the amplitude
and coherence image are exported as floating point values (unless
mapped to byte by the user), while the interferometric phase is stored
as a color-coded byte image. Still need to work out how to deal with
exporting the interferogram as complex signal.