Some more netCDF and now HDF5 goodness.
The HDF part is a preliminary check in.

For the netCDF part I enhanced the resulting file by adding three
groups of metadata parameters: overview, SAR and orbit. I only added
the real essential metadata. Some of the parameters required the
definition of units to be meaningful. This is where things get fancy
in the implementaton as it required the definition of ten separate
compound types (one for each unit I used in the metadata).

The netCDF export also calls now 'meta_write_xml' to generate an XML
version the ASF internal metadata that accompanies the resulting .nc file.