Added sinusoidal map projection to the mix.
Ran into an exception case during the testing. The sinusoidal map projection (as it turned out the equi-rectangular and Mercator as well) is not well defined when the image is crossing the dateline. Did some tests with gdal to confirm this (had the same problems). Bottom line is that I put a check in place whether the image is crossing the dateline for these three map projections. In case the respective image does the function errors out. Probably the cleanest solution.

Now we should have the functionality we need for the terrestrial ecology part of the HDF project.