Additional changes for darwin compatibility.
Upgraded the packaged gsl to 1.15 because 1.6 was failing to build on darwin.

Some header files where including malloc.h which is not provided by darwin, used stdlib.h instead.

Fixed a problem where closedir could be called on a bad dirhandle, causing a segfault on darwin.

The parse_val function in parse_option.c used to use g_strdup for string duplication. For an unknown reason, that function was returning an out-of-range
pointer for me. Since it did not look like using glib function was necessary there I changed the calls to use STRDUP and equivalent functions.