Adapted the radiometric terrain correction for amplitude images.
The tool now accepts only amplitude images and radiometrically corrects the amplitude values. For the radiometric correction I implemented a function 'get_rad_cal_dn' that takes care of all calibration schemes.

Plenty of testing to be done to verify that these calibration schemes actually all work.

The calibration parameters are now applied with the new 'asf_calibrate' tool. Working on that next.