[TA3029] Implemented asf_calibrate tool.
Currently, it requires amplitude images are input. The functionality for "remapping" different radiometries is developed but commented out for the moment. Don't have enough time for alpha testing.

The hooks for Woods Hole mapping to byte are in the wrapper tool. Still need to be flashed out in the library function. I have coded up most of it already but never checked it in.