[TA3032] Fixed asf_mapready (including GUI) to handle end to end processing.
Realized that I hadn't hooked 'make_gr_dem' to the processing flow, required to generated the required input for 'rtc'. That is what we are going with for the moment.

Once Eddie's changes to asf_terrcorr include all bells and whistles, we need to adjust a few things in the library code.