[TA3041] Implemented the Woods Hole style scaling to byte.
Scaling for single and quad-pol data is done per band. Dual-pol data gets a special treatment. As a third band the difference of band 1 and band 2 is calculated. This three-band image is always exported as RGB.

In order to pass this RGB combination appropriately to asf_export, I defined a new image data type 'RGB_STACK'.