Monster check in because I could not split things up easily. Let's see what we got.
Added detected imagery to the TerraSAR-X ingest. Only needed a few changes for that. The addition of the calibration for TerraSAR-X took a little more doing. So far we only applied calibration parameters to CEOS data (and TerraSAR-X data comes as XML metadata). For that I needed to move the XML support from the libasf_import to asf_meta library. That required a bunch of changes in Makefiles to accommodate that change.

I enhanced the SMAP ingest so that we can generate subsets based on geographic coordinates. I added the geographic map projection to the list (was much easier to evaluate that way). Some initial tests showed that geocoding of entire SMAP data sets without using the pre-calculated EASE grid indices is probably not feasible. Will need to add a few checks and balances to the current functionality to prevent the user from doing some time intensive stuff.