Since we are moving things to github, I wanted to check in some work in progress for SMAP.
This is mostly related to the geocoding of SMAP data in their native EASE grid coordinates. The current implementation includes
- definition of EASE grid global, EASE grid north and EASE grid south map projection
- reading/writing of metadata that contains this new map projections
- exporting these map projections into GeoTIFF using all proper GeoTIFF EPSG tags

What is missing right now is the actual geocoding using the grid indices provided in the original SMAP HDF5 data. That will require a little more plumbing to establish the mapping function. Whenever I get to that.