Added -input, -output and -tmpdir options to asf_mapready.
This allows to either (temporarily) overwrite the parameters in a configuration file or the use of a "settings configuration file" that has these parameters missing. This way we can define a settings configuration file for a particular test case and run the same test case with various data sets.

In order to get this completely to work, we still need to fix DE719, as currently there is some mismatch of the contents in configuration files.