[DE768]
Quite a number of things need to be changed to make polarimetric processing work within MapReady when the data in radiometrically terrain corrected as well.

Kirk actually checked the first fix in. Radiometric terrain needs to be only applied to amplitude bands, as we are only trying to adjust the brightness of pixel values. Hence, no change in phase values are required.

The polarimetric processing (except Faraday rotation) needed to be shifted in the processing flow, as it is applied to sigma power scale values in the first place. That means for our current scheme that this has to happen after terrain correction and applying the calibration parameters.

The calibration step had the same issue as the radiometric correction. The calibration parameters only need to be applied to amplitude bands (equivalent to applying the calibration to I and Q in the complex world).

Finally, the new geocoding scheme of ALOS L1.1 data (using a slant range transformation block in ground range) did not account for multilooking of data (happens by ensemble averaging during the polarimetric processing).