Added a MapReady settings configuration file.
That takes care of most of the parameters that can be passed to MapReady via configuration file or command line but are not available within the GUI (especially default terrain correction settings from the bake off).

This change allows the user to specify the location of ERS precision state vectors. However, I am pretty sure we can't easily mix E1 and E2 state vectors.

Another nice one is the definition of a standard data input directory. Would be nice if the Windows version of MapReady could now pick up that one.