New tool: add_aux_band
Adds another band to the input file with incidence angles or doppler
or whatever the user requests on the command line.  For example:
  add_aux_band incidence_angle in out

This is for JPL, who needs to be able to add incidence angles as a
second band before geocoding in order to get those values without
having to perform terrain correction.