Fix issues in debian package.
The issues fixed include the lack of a copyright file, the lack of a
changelog, the unstripped binaries, and the lack of an installed size in
the control file. The installed size is generated from the size of the
files in the asf-mapready subdirectory, the copyright file was written
from scratch, and the changelog is generated from the last 10 commits.
The only problem with that is that any user that wants to create this
package must have git installed.