Use pkg-config to search for libraries.
The PKG_CHECK_MODULE macro is now used for each library that should be
linked against. pkg-config files do not exist for some libraries, but I
included the macro anyway, in case those files are added in future
releases. In the event that the library is not found by pkg-config,
there could be smarter code to locate it.