Add a macro to hunt for header files.
This macro uses find to search through /usr/include for a header file
that should be included with a library. If it finds it, it sets the
appropriate CFLAGS variable, and if it doesn't, it exits with an error.
It might not be the most appealing solution, but I think it's the most
portable.